카페 메뉴 저장 수정

+++ src/main/java/com/finalProject/Back/repository/CafeMapper.java
@@ -13,7 +13,7 @@ import java.util.List;
 @Mapper
 public interface CafeMapper{
     int save(Cafe cafe);
-    int saveDefaultCafeMenu(int cafeId);
+    int saveDefaultCafeMenu(Long cafeId);
     Cafe findByUserId(Long id);
     CafeDetail findByCafeId(Long cafeId);
     List<Cafe> findBySearchAndCategory(@Param("category") String category, @Param("search") String search);
